Allow team rotation to be float
The CDS only allowed the team's desk rotation to be integers, but we made the spec support floating point values. Updating the contest model to support double, and getting rid of float/short in FloorMap utility and downstream effect on FloorGenerators and BalloonPrinter. Using this opportunity to check in the NAC floor map gen too.

+++ BalloonUtility/src/org/icpc/tools/balloon/BalloonPrinter.java
@@ -613,11 +613,11 @@ public class BalloonPrinter {
 			float yt = r.y + (int) ((team.getY() - bounds.y) * scale);
 
 			// team area
-			float teamAreaDepth = floor.teamAreaDepth;
-			float teamAreaWidth = floor.teamAreaWidth;
+			double teamAreaDepth = floor.teamAreaDepth;
+			double teamAreaWidth = floor.teamAreaWidth;
 			float dx = 1f;
 			transform.translate(xt, yt);
-			transform.rotate(-team.getRotation());
+			transform.rotate((float) -team.getRotation());
 			gc.setTransform(transform);
 
 			String id = team.getId();
@@ -631,8 +631,8 @@ public class BalloonPrinter {
 			gc.drawRectangle(tr1);
 
 			// chairs
-			float tableDepth = floor.tableDepth;
-			float tableWidth = floor.tableWidth;
+			double tableDepth = floor.tableDepth;
+			double tableWidth = floor.tableWidth;
 
 			// table
 			Rectangle tr = new Rectangle((int) (-tableDepth * scale / 2f), (int) (-tableWidth * scale / 2f),
